requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-api-python-client==2.107.0
google-generativeai==0.3.1
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.2
beautifulsoup4==4.12.2
twilio==8.10.0